// frontend/app/src/app/core/models/treatment.model.ts
export interface treatment {
    id: number;
    name: string;
    medicine: string;
    dose: string;
    frequency: string;
    instructions: string;
    startDate: string;
    endDate: string;
    veterinarianFullName: string;
    duration: string;
    status: string;
    suspendedReason: string;
}